Alter whitespace to fix markup, and improve readability
Exercise numbers were displaying as 1. 1. 1. instead in incrementing.
..because the inner paragraphs were not indented.
..AND because there were 2 line breaks before each new line number,
instead of the required 1.
This has been fixed.

"Entering Nightmare Mode" also needed to be indented
  to continue proper line numbering following this heading.
It is now indented.
An h3 tag was added, to indicate this as a separate section, as intended.

h1 tag now displays as intended.
The fix was to add a space after the hash.

Long sentences were broken into separate lines by adding 2 blank spaces followed by a CR.
The double space at the end of a line, allows the text to flow onto a new line,
without starting a new paragraph.